Add language to local queries.
